Revert node path fix by fixing root cause which was failure to update PATH
