+++ .pax/prepare-workspace.sh
@@ -120,6 +120,8 @@ mkdir -p "${CONTENT_DIR}/schemas"
 # Building TS files
 cd "${ROOT_DIR}/build/zwe"
 npm ci && npm run prod
+# Cleanup TS files after build
+rm -f "${ROOT_DIR}/bin/**/*.ts"
 
 
 # copy from current github source
