+++ bin/libs/node.ts
@@ -31,9 +31,9 @@ std.setenv('__UNTAGGED_READ_MODE','V6');
 
 
 export function ensureNodeIsOnPath(): void {
-  let path=std.getenv('PATH');
+  let path=std.getenv('PATH') || '';
   let nodeHome=std.getenv('NODE_HOME');
-  if (path && !path.includes(`:${nodeHome}/bin:`)) {
+  if (!path.includes(`:${nodeHome}/bin:`)) {
     std.setenv('PATH', `${nodeHome}/bin:${path}`);
   }
 }
