+++ bin/libs/shell.ts
@@ -15,6 +15,7 @@ import * as xplatform from 'xplatform';
 
 import * as fs from './fs';
 import * as stringlib from './string';
+import * as bufferlib from './buffer';
 
 declare namespace console {
   function log(...args:string[]): void;
@@ -31,57 +32,6 @@ export type ExecReturn = {
 
 export type path = string;
 
-const hexDigits = [ "0", "1", "2", "3",
-                    "4", "5", "6", "7",
-                    "8", "9", "A", "B",
-                    "C", "D", "E", "F" ];
-
-class ExpandableBuffer {
-  size:number;
-  buffer:ArrayBuffer;
-  bytes:Uint8Array;
-  pos:number;
-  
-  constructor(size:number){
-    this.size = size;
-    this.buffer = new ArrayBuffer(size);
-    this.bytes = new Uint8Array(this.buffer);
-    this.pos = 0;
-  }
-
-  append(data:Uint8Array, offset:number, length:number):void {
-    if (this.pos+length > this.size){
-      let newSize = this.size * 2;
-      let newBuffer = new ArrayBuffer(newSize);
-      let newBytes = new Uint8Array(newBuffer);
-      newBytes.set(this.bytes,0);
-      this.bytes = newBytes;
-      this.buffer = newBuffer;
-      this.size = newSize;
-    }
-    this.bytes.set(data.slice(offset,length), this.pos);
-    this.pos += length;
-  }
-
-  dump(lim:number):string {
-    let chunks = [];
-    for (let i=0; i<lim; i++){
-      let b = this.bytes[i];
-      let h = (b >> 4)&0xf;
-      let l = b&0xf;
-      chunks.push(hexDigits[h]+hexDigits[l]+"");
-      if ((i % 4) == 3){ chunks.push(" "); }
-      if ((i % 16) == 15){ chunks.push("\n"); }
-    }
-    return chunks.join("");
-  }
-
-  getString():string{
-    return xplatform.stringFromBytes(this.buffer, 0, this.pos, xplatform.AUTO_DETECT);
-  }
-}
-
-
 export function execSync(command: string, ...args: string[]): ExecReturn {
   const rc = os.exec([command, ...args],
                      {block: true, usePath: true});
@@ -92,7 +42,7 @@ export function execSync(command: string, ...args: string[]): ExecReturn {
 
 function readStreamFully(fd:number):string{
   let readBuffer = new Uint8Array(BUFFER_SIZE);
-  let fileBuffer = new ExpandableBuffer(BUFFER_SIZE);
+  let fileBuffer = new bufferlib.ExpandableBuffer(BUFFER_SIZE);
   
   let bytesRead = 0;
   do {
